+++ wasm_csv_project/csv_converter.js
@@ -2,13 +2,13 @@ var l;l||=typeof Module != 'undefined' ? Module : {};var aa="./this.program",ba=
 function la(){var a=F.buffer;t=new Int8Array(a);A=new Int16Array(a);v=new Uint8Array(a);B=new Uint16Array(a);C=new Int32Array(a);D=new Uint32Array(a);ha=new Float32Array(a);ia=new Float64Array(a);ja=new BigInt64Array(a);ka=new BigUint64Array(a)}function ma(a){l.onAbort?.(a);a="Aborted("+a+")";p(a);r=!0;throw new WebAssembly.RuntimeError(a+". Build with -sASSERTIONS for more info.");}var na;
 async function oa(a){if(!q)try{var b=await ea(a);return new Uint8Array(b)}catch{}if(a==na&&q)a=new Uint8Array(q);else throw"both async and sync fetching of the wasm failed";return a}async function pa(a,b){try{var c=await oa(a);return await WebAssembly.instantiate(c,b)}catch(d){p(`failed to asynchronously prepare wasm: ${d}`),ma(d)}}
 async function qa(a){var b=na;if(!q)try{var c=fetch(b,{credentials:"same-origin"});return await WebAssembly.instantiateStreaming(c,a)}catch(d){p(`wasm streaming compile failed: ${d}`),p("falling back to ArrayBuffer instantiation")}return pa(b,a)}class sa{name="ExitStatus";constructor(a){this.message=`Program terminated with exit(${a})`;this.status=a}}var ta=a=>{for(;0<a.length;)a.shift()(l)},ua=[],va=[],wa=()=>{var a=l.preRun.shift();va.push(a)},G=0,H=null,I=!0,xa=[],ya=0,J=0;
-class za{constructor(a){this.oa=a;this.ia=a-24}}var Ba=a=>{var b=J;if(!b)return L(0),0;var c=new za(b);D[c.ia+16>>>2>>>0]=b;var d=D[c.ia+4>>>2>>>0];if(!d)return L(0),b;for(var e of a){if(0===e||e===d)break;if(Aa(e,d,c.ia+16))return L(e),b}L(d);return b},M=a=>{a>>>=0;for(var b="";;){var c=v[a++>>>0];if(!c)return b;b+=String.fromCharCode(c)}},N={},O={},Ca={},P=class extends Error{constructor(a){super(a);this.name="BindingError"}},Da=a=>{throw new P(a);};
-function Ea(a,b,c={}){var d=b.name;if(!a)throw new P(`type "${d}" must have a positive integer typeid pointer`);if(O.hasOwnProperty(a)){if(c.pa)return;throw new P(`Cannot register type '${d}' twice`);}O[a]=b;delete Ca[a];N.hasOwnProperty(a)&&(b=N[a],delete N[a],b.forEach(e=>e()))}function Q(a,b,c={}){return Ea(a,b,c)}
+class za{constructor(a){this.qa=a;this.ka=a-24}}var Ba=a=>{var b=J;if(!b)return L(0),0;var c=new za(b);D[c.ka+16>>>2>>>0]=b;var d=D[c.ka+4>>>2>>>0];if(!d)return L(0),b;for(var e of a){if(0===e||e===d)break;if(Aa(e,d,c.ka+16))return L(e),b}L(d);return b},M=a=>{a>>>=0;for(var b="";;){var c=v[a++>>>0];if(!c)return b;b+=String.fromCharCode(c)}},N={},O={},Ca={},P=class extends Error{constructor(a){super(a);this.name="BindingError"}},Da=a=>{throw new P(a);};
+function Ea(a,b,c={}){var d=b.name;if(!a)throw new P(`type "${d}" must have a positive integer typeid pointer`);if(O.hasOwnProperty(a)){if(c.ra)return;throw new P(`Cannot register type '${d}' twice`);}O[a]=b;delete Ca[a];N.hasOwnProperty(a)&&(b=N[a],delete N[a],b.forEach(e=>e()))}function Q(a,b,c={}){return Ea(a,b,c)}
 var Fa=(a,b,c)=>{switch(b){case 1:return c?d=>t[d>>>0]:d=>v[d>>>0];case 2:return c?d=>A[d>>>1>>>0]:d=>B[d>>>1>>>0];case 4:return c?d=>C[d>>>2>>>0]:d=>D[d>>>2>>>0];case 8:return c?d=>ja[d>>>3>>>0]:d=>ka[d>>>3>>>0];default:throw new TypeError(`invalid integer width (${b}): ${a}`);}},Ga=[],R=[0,1,,1,null,1,!0,1,!1,1],Ha=a=>{switch(a){case void 0:return 2;case null:return 4;case !0:return 6;case !1:return 8;default:const b=Ga.pop()||R.length;R[b]=a;R[b+1]=1;return b}};
-function Ia(a){return this.ha(D[a>>>2>>>0])}var Ja={name:"emscripten::val",ha:a=>{if(!a)throw new P(`Cannot use deleted val. handle = ${a}`);var b=R[a];a>>>=0;9<a&&0===--R[a+1]&&(R[a]=void 0,Ga.push(a));return b},la:(a,b)=>Ha(b),ma:Ia,ka:null},Ka=(a,b)=>{switch(b){case 4:return function(c){return this.ha(ha[c>>>2>>>0])};case 8:return function(c){return this.ha(ia[c>>>3>>>0])};default:throw new TypeError(`invalid float width (${b}): ${a}`);}},La=a=>{for(;a.length;){var b=a.pop();a.pop()(b)}};
-function Ma(a){for(var b=1;b<a.length;++b)if(null!==a[b]&&void 0===a[b].ka)return!0;return!1}
-var Na=(a,b)=>{var c=l;if(void 0===c[a].ja){var d=c[a];c[a]=function(...e){if(!c[a].ja.hasOwnProperty(e.length))throw new P(`Function '${b}' called with an invalid number of arguments (${e.length}) - expects one of (${c[a].ja})!`);return c[a].ja[e.length].apply(this,e)};c[a].ja=[];c[a].ja[d.na]=d}},Oa=(a,b,c)=>{if(l.hasOwnProperty(a)){if(void 0===c||void 0!==l[a].ja&&void 0!==l[a].ja[c])throw new P(`Cannot register public name '${a}' twice`);Na(a,a);if(l[a].ja.hasOwnProperty(c))throw new P(`Cannot register multiple overloads of a function with the same number of arguments (${c})!`);
-l[a].ja[c]=b}else l[a]=b,l[a].na=c},Pa=(a,b)=>{for(var c=[],d=0;d<a;d++)c.push(D[b+4*d>>>2>>>0]);return c},Qa=class extends Error{constructor(a){super(a);this.name="InternalError"}},Ra=[],S=a=>{var b=Ra[a];b||(Ra[a]=b=Sa.get(a));return b},Ta=(a,b,c=[])=>{b=S(b)(...c);return"p"==a[0]?b>>>0:b},Ua=(a,b)=>(...c)=>Ta(a,b,c),Va=(a,b)=>{a=M(a);var c=a.includes("p")?Ua(a,b):S(b);if("function"!=typeof c)throw new P(`unknown function pointer with signature ${a}: ${b}`);return c};class Wa extends Error{}
+function Ia(a){return this.ja(D[a>>>2>>>0])}var Ja={name:"emscripten::val",ja:a=>{if(!a)throw new P(`Cannot use deleted val. handle = ${a}`);var b=R[a];a>>>=0;9<a&&0===--R[a+1]&&(R[a]=void 0,Ga.push(a));return b},na:(a,b)=>Ha(b),oa:Ia,ma:null},Ka=(a,b)=>{switch(b){case 4:return function(c){return this.ja(ha[c>>>2>>>0])};case 8:return function(c){return this.ja(ia[c>>>3>>>0])};default:throw new TypeError(`invalid float width (${b}): ${a}`);}},La=a=>{for(;a.length;){var b=a.pop();a.pop()(b)}};
+function Ma(a){for(var b=1;b<a.length;++b)if(null!==a[b]&&void 0===a[b].ma)return!0;return!1}
+var Na=(a,b)=>{var c=l;if(void 0===c[a].la){var d=c[a];c[a]=function(...e){if(!c[a].la.hasOwnProperty(e.length))throw new P(`Function '${b}' called with an invalid number of arguments (${e.length}) - expects one of (${c[a].la})!`);return c[a].la[e.length].apply(this,e)};c[a].la=[];c[a].la[d.pa]=d}},Oa=(a,b,c)=>{if(l.hasOwnProperty(a)){if(void 0===c||void 0!==l[a].la&&void 0!==l[a].la[c])throw new P(`Cannot register public name '${a}' twice`);Na(a,a);if(l[a].la.hasOwnProperty(c))throw new P(`Cannot register multiple overloads of a function with the same number of arguments (${c})!`);
+l[a].la[c]=b}else l[a]=b,l[a].pa=c},Pa=(a,b)=>{for(var c=[],d=0;d<a;d++)c.push(D[b+4*d>>>2>>>0]);return c},Qa=class extends Error{constructor(a){super(a);this.name="InternalError"}},Ra=[],S=a=>{var b=Ra[a];b||(Ra[a]=b=Sa.get(a));return b},Ta=(a,b,c=[])=>{b=S(b)(...c);return"p"==a[0]?b>>>0:b},Ua=(a,b)=>(...c)=>Ta(a,b,c),Va=(a,b)=>{a=M(a);var c=a.includes("p")?Ua(a,b):S(b);if("function"!=typeof c)throw new P(`unknown function pointer with signature ${a}: ${b}`);return c};class Wa extends Error{}
 var Ya=a=>{a=Xa(a);var b=M(a);T(a);return b},ab=(a,b)=>{function c(f){e[f]||O[f]||(Ca[f]?Ca[f].forEach(c):(d.push(f),e[f]=!0))}var d=[],e={};b.forEach(c);throw new Wa(`${a}: `+d.map(Ya).join([", "]));},bb=(a,b)=>{function c(h){h=b(h);if(h.length!==d.length)throw new Qa("Mismatched type converter count");for(var k=0;k<d.length;++k)Q(d[k],h[k])}var d=[];d.forEach(h=>Ca[h]=a);var e=Array(a.length),f=[],g=0;for(let [h,k]of a.entries())O.hasOwnProperty(k)?e[h]=O[k]:(f.push(k),N.hasOwnProperty(k)||(N[k]=
 []),N[k].push(()=>{e[h]=O[k];++g;g===f.length&&c(e)}));0===f.length&&c(e)},cb=a=>{a=a.trim();const b=a.indexOf("(");return-1===b?a:a.slice(0,b)},U=(a,b,c)=>{var d=v;b>>>=0;if(!(0<c))return 0;var e=b;c=b+c-1;for(var f=0;f<a.length;++f){var g=a.codePointAt(f);if(127>=g){if(b>=c)break;d[b++>>>0]=g}else if(2047>=g){if(b+1>=c)break;d[b++>>>0]=192|g>>6;d[b++>>>0]=128|g&63}else if(65535>=g){if(b+2>=c)break;d[b++>>>0]=224|g>>12;d[b++>>>0]=128|g>>6&63;d[b++>>>0]=128|g&63}else{if(b+3>=c)break;d[b++>>>0]=240|
 g>>18;d[b++>>>0]=128|g>>12&63;d[b++>>>0]=128|g>>6&63;d[b++>>>0]=128|g&63;f++}}d[b>>>0]=0;return b-e},db=a=>{for(var b=0,c=0;c<a.length;++c){var d=a.charCodeAt(c);127>=d?b++:2047>=d?b+=2:55296<=d&&57343>=d?(b+=4,++c):b+=3}return b},eb=globalThis.TextDecoder&&new TextDecoder,fb=(a,b,c,d)=>{c=b+c;if(d)return c;for(;a[b]&&!(b>=c);)++b;return b},gb=(a=0,b,c)=>{var d=v;a>>>=0;b=fb(d,a,b,c);if(16<b-a&&d.buffer&&eb)return eb.decode(d.subarray(a,b));for(c="";a<b;){var e=d[a++];if(e&128){var f=d[a++]&63;if(192==
@@ -17,20 +17,20 @@ if(2>c)return 0;c-=2;var d=b;c=c<2*a.length?c/2:a.length;for(var e=0;e<c;++e)A[b
 a.codePointAt(c)&&c++,b+=4;return b},ob=0,V={},pb=a=>{if(!(a instanceof sa||"unwind"==a))throw a;},qb=a=>{fa=a;I||0<ob||(l.onExit?.(a),r=!0);throw new sa(a);},rb=a=>{if(!r)try{if(a(),!(I||0<ob))try{fa=a=fa,qb(a)}catch(b){pb(b)}}catch(b){pb(b)}},sb={},ub=()=>{if(!tb){var a={USER:"web_user",LOGNAME:"web_user",PATH:"/",PWD:"/",HOME:"/home/web_user",LANG:("object"==typeof navigator&&navigator.language||"C").replace("-","_")+".UTF-8",_:aa||"./this.program"},b;for(b in sb)void 0===sb[b]?delete a[b]:a[b]=
 sb[b];var c=[];for(b in a)c.push(`${b}=${a[b]}`);tb=c}return tb},tb,wb=(a,b,c,d)=>{var e={string:m=>{var w=0;if(null!==m&&void 0!==m&&0!==m){w=db(m)+1;var z=vb(w);U(m,z,w);w=z}return w},array:m=>{var w=vb(m.length);t.set(m,w>>>0);return w}};a=l["_"+a];var f=[],g=0;if(d)for(var h=0;h<d.length;h++){var k=e[c[h]];k?(0===g&&(g=W()),f[h]=k(d[h])):f[h]=d[h]}c=a(...f);return c=function(m){0!==g&&X(g);m="string"===b?(m>>>=0)?gb(m):"":"boolean"===b?!!m:m;return m}(c)};l.noExitRuntime&&(I=l.noExitRuntime);
 l.printErr&&(p=l.printErr);l.wasmBinary&&(q=l.wasmBinary);l.thisProgram&&(aa=l.thisProgram);if(l.preInit)for("function"==typeof l.preInit&&(l.preInit=[l.preInit]);0<l.preInit.length;)l.preInit.shift()();l.ccall=wb;l.cwrap=(a,b,c,d)=>{var e=!c||c.every(f=>"number"===f||"boolean"===f);return"string"!==b&&e&&!d?l["_"+a]:(...f)=>wb(a,b,c,f,d)};
-var xb,Xa,yb,T,Y,L,X,vb,W,zb,Ab,Aa,Bb,F,Sa,Vb={q:function(a){a>>>=0;var b=new za(a);0==t[b.ia+12>>>0]&&(t[b.ia+12>>>0]=1,ya--);t[b.ia+13>>>0]=0;xa.push(b);Ab(a);return Bb(a)},s:()=>{Y(0,0);var a=xa.pop();zb(a.oa);J=0},a:function(){return Ba([])},j:function(a){return Ba([a>>>0])},R:()=>{var a=xa.pop();a||ma("no exception to throw");var b=a.oa;0==t[a.ia+13>>>0]&&(xa.push(a),t[a.ia+13>>>0]=1,t[a.ia+12>>>0]=0,ya++);J=b;throw J;},k:function(a,b,c){a>>>=0;var d=new za(a);D[d.ia+16>>>2>>>0]=0;D[d.ia+4>>>
-2>>>0]=b>>>0;D[d.ia+8>>>2>>>0]=c>>>0;J=a;ya++;throw J;},Q:()=>ya,d:function(a){J||=a>>>0;throw J;},M:()=>ma(""),C:function(a,b,c,d,e){a>>>=0;c>>>=0;b=M(b>>>0);d=0n===d;let f=g=>g;if(d){const g=8*c;f=h=>BigInt.asUintN(g,h);e=f(e)}Q(a,{name:b,ha:f,la:(g,h)=>{"number"==typeof h&&(h=BigInt(h));return h},ma:Fa(b,c,!d),ka:null})},F:function(a,b,c,d){b=M(b>>>0);Q(a>>>0,{name:b,ha:function(e){return!!e},la:function(e,f){return f?c:d},ma:function(e){return this.ha(v[e>>>0])},ka:null})},D:function(a){return Q(a>>>
-0,Ja)},B:function(a,b,c){b=M(b>>>0);Q(a>>>0,{name:b,ha:d=>d,la:(d,e)=>e,ma:Ka(b,c>>>0),ka:null})},t:function(a,b,c,d,e,f,g){a>>>=0;d>>>=0;e>>>=0;f>>>=0;var h=Pa(b,c>>>0);a=M(a);a=cb(a);e=Va(d,e);Oa(a,function(){ab(`Cannot call ${a} due to unbound types`,h)},b-1);bb(h,k=>{var m=a;var w=a;k=[k[0],null].concat(k.slice(1));var z=e,u=f,n=k.length;if(2>n)throw new P("argTypes array size mismatch! Must at least get return value and 'this' types!");var E=null!==k[1]&&!1,K=Ma(k),da=!k[0].qa,y=k[0],x=k[1];
-z=[w,Da,z,u,La,y.ha.bind(y),x?.la.bind(x)];for(u=2;u<n;++u)y=k[u],z.push(y.la.bind(y));if(!K)for(u=E?1:2;u<k.length;++u)null!==k[u].ka&&z.push(k[u].ka);K=Ma(k);u=k.length-2;x=[];y=["fn"];E&&y.push("thisWired");for(n=0;n<u;++n)x.push(`arg${n}`),y.push(`arg${n}Wired`);x=x.join(",");y=y.join(",");x=`return function (${x}) {\n`;K&&(x+="var destructors = [];\n");var Za=K?"destructors":"null",ra="humanName throwBindingError invoker fn runDestructors fromRetWire toClassParamWire".split(" ");E&&(x+=`var thisWired = toClassParamWire(${Za}, this);\n`);
-for(n=0;n<u;++n){var $a=`toArg${n}Wire`;x+=`var arg${n}Wired = ${$a}(${Za}, arg${n});\n`;ra.push($a)}x+=(da||g?"var rv = ":"")+`invoker(${y});\n`;if(K)x+="runDestructors(destructors);\n";else for(n=E?1:2;n<k.length;++n)E=1===n?"thisWired":"arg"+(n-2)+"Wired",null!==k[n].ka&&(x+=`${E}_dtor(${E});\n`,ra.push(`${E}_dtor`));da&&(x+="var ret = fromRetWire(rv);\nreturn ret;\n");k=(new Function(ra,x+"}\n"))(...z);w=Object.defineProperty(k,"name",{value:w});k=b-1;if(!l.hasOwnProperty(m))throw new Qa("Replacing nonexistent public symbol");
-void 0!==l[m].ja&&void 0!==k?l[m].ja[k]=w:(l[m]=w,l[m].na=k);return[]})},r:function(a,b,c,d,e){a>>>=0;c>>>=0;b=M(b>>>0);let f=h=>h;if(0===d){var g=32-8*c;f=h=>h<<g>>>g;e=f(e)}Q(a,{name:b,ha:f,la:(h,k)=>k,ma:Fa(b,c,0!==d),ka:null})},m:function(a,b,c){function d(f){return new e(t.buffer,D[f+4>>>2>>>0],D[f>>>2>>>0])}var e=[Int8Array,Uint8Array,Int16Array,Uint16Array,Int32Array,Uint32Array,Float32Array,Float64Array,BigInt64Array,BigUint64Array][b];c=M(c>>>0);Q(a>>>0,{name:c,ha:d,ma:d},{pa:!0})},E:function(a,
-b){b=M(b>>>0);Q(a>>>0,{name:b,ha(c){var d=(d=c+4>>>0)?gb(d,D[c>>>2>>>0],!0):"";T(c);return d},la(c,d){d instanceof ArrayBuffer&&(d=new Uint8Array(d));var e="string"==typeof d;if(!(e||ArrayBuffer.isView(d)&&1==d.BYTES_PER_ELEMENT))throw new P("Cannot pass non-string to std::string");var f=e?db(d):d.length;var g=xb(4+f+1),h=g+4;D[g>>>2>>>0]=f;e?U(d,h,f+1):v.set(d,h>>>0);null!==c&&c.push(T,g);return g},ma:Ia,ka(c){T(c)}})},y:function(a,b,c){b>>>=0;c>>>=0;c=M(c);if(2===b){var d=ib;var e=jb;var f=kb}else d=
-lb,e=mb,f=nb;Q(a>>>0,{name:c,ha:g=>{var h=d(g+4,D[g>>>2>>>0]*b,!0);T(g);return h},la:(g,h)=>{if("string"!=typeof h)throw new P(`Cannot pass non-string to C++ string type ${c}`);var k=f(h),m=xb(4+k+b);D[m>>>2>>>0]=k/b;e(h,m+4,k+b);null!==g&&g.push(T,m);return m},ma:Ia,ka(g){T(g)}})},G:function(a,b){b=M(b>>>0);Q(a>>>0,{qa:!0,name:b,ha:()=>{},la:()=>{}})},K:()=>{I=!1;ob=0},L:(a,b)=>{V[a]&&(clearTimeout(V[a].id),delete V[a]);if(!b)return 0;var c=setTimeout(()=>{delete V[a];rb(()=>yb(a,performance.now()))},
-b);V[a]={id:c,ra:b};return 0},N:function(a,b,c,d){c>>>=0;d>>>=0;var e=(new Date).getFullYear(),f=(new Date(e,0,1)).getTimezoneOffset();e=(new Date(e,6,1)).getTimezoneOffset();D[a>>>0>>>2>>>0]=60*Math.max(f,e);C[b>>>0>>>2>>>0]=Number(f!=e);b=g=>{var h=Math.abs(g);return`UTC${0<=g?"-":"+"}${String(Math.floor(h/60)).padStart(2,"0")}${String(h%60).padStart(2,"0")}`};a=b(f);b=b(e);e<f?(U(a,c,17),U(b,d,17)):(U(a,d,17),U(b,c,17))},S:function(a){a>>>=0;var b=v.length;if(4294901760<a)return!1;for(var c=1;4>=
-c;c*=2){var d=b*(1+.2/c);d=Math.min(d,a+100663296);a:{d=(Math.min(4294901760,65536*Math.ceil(Math.max(a,d)/65536))-F.buffer.byteLength+65535)/65536|0;try{F.grow(d);la();var e=1;break a}catch(f){}e=void 0}if(e)return!0}return!1},O:function(a,b){a>>>=0;b>>>=0;var c=0,d=0,e;for(e of ub()){var f=b+c;D[a+d>>>2>>>0]=f;c+=U(e,f,Infinity)+1;d+=4}return 0},P:function(a,b){a>>>=0;b>>>=0;var c=ub();D[a>>>2>>>0]=c.length;a=0;for(var d of c)a+=db(d)+1;D[b>>>2>>>0]=a;return 0},b:Cb,v:Db,c:Eb,g:Fb,z:Gb,I:Hb,p:Ib,
-l:Jb,A:Kb,x:Lb,H:Mb,f:Nb,o:Ob,e:Pb,h:Qb,i:Rb,n:Sb,u:Tb,w:Ub,J:qb};function Nb(a){var b=W();try{S(a)()}catch(c){X(b);if(c!==c+0)throw c;Y(1,0)}}function Cb(a,b){var c=W();try{return S(a)(b)}catch(d){X(c);if(d!==d+0)throw d;Y(1,0)}}function Qb(a,b,c,d){var e=W();try{S(a)(b,c,d)}catch(f){X(e);if(f!==f+0)throw f;Y(1,0)}}function Pb(a,b,c){var d=W();try{S(a)(b,c)}catch(e){X(d);if(e!==e+0)throw e;Y(1,0)}}function Rb(a,b,c,d,e){var f=W();try{S(a)(b,c,d,e)}catch(g){X(f);if(g!==g+0)throw g;Y(1,0)}}
-function Fb(a,b,c,d){var e=W();try{return S(a)(b,c,d)}catch(f){X(e);if(f!==f+0)throw f;Y(1,0)}}function Eb(a,b,c){var d=W();try{return S(a)(b,c)}catch(e){X(d);if(e!==e+0)throw e;Y(1,0)}}function Jb(a,b,c,d,e,f,g){var h=W();try{return S(a)(b,c,d,e,f,g)}catch(k){X(h);if(k!==k+0)throw k;Y(1,0)}}function Ib(a,b,c,d,e,f){var g=W();try{return S(a)(b,c,d,e,f)}catch(h){X(g);if(h!==h+0)throw h;Y(1,0)}}function Db(a,b,c){var d=W();try{return S(a)(b,c)}catch(e){X(d);if(e!==e+0)throw e;Y(1,0)}}
-function Ob(a,b){var c=W();try{S(a)(b)}catch(d){X(c);if(d!==d+0)throw d;Y(1,0)}}function Gb(a,b,c,d,e){var f=W();try{return S(a)(b,c,d,e)}catch(g){X(f);if(g!==g+0)throw g;Y(1,0)}}function Hb(a,b,c,d,e,f){var g=W();try{return S(a)(b,c,d,e,f)}catch(h){X(g);if(h!==h+0)throw h;Y(1,0)}}function Mb(a,b,c){var d=W();try{return S(a)(b,c)}catch(e){X(d);if(e!==e+0)throw e;Y(1,0)}}function Kb(a,b,c,d,e,f,g,h){var k=W();try{return S(a)(b,c,d,e,f,g,h)}catch(m){X(k);if(m!==m+0)throw m;Y(1,0)}}
-function Sb(a,b,c,d,e,f,g,h){var k=W();try{S(a)(b,c,d,e,f,g,h)}catch(m){X(k);if(m!==m+0)throw m;Y(1,0)}}function Lb(a,b,c,d,e,f,g,h,k,m,w,z){var u=W();try{return S(a)(b,c,d,e,f,g,h,k,m,w,z)}catch(n){X(u);if(n!==n+0)throw n;Y(1,0)}}function Tb(a,b,c,d,e,f,g,h,k,m,w){var z=W();try{S(a)(b,c,d,e,f,g,h,k,m,w)}catch(u){X(z);if(u!==u+0)throw u;Y(1,0)}}function Ub(a,b,c,d,e,f,g,h,k,m,w,z,u,n,E,K){var da=W();try{S(a)(b,c,d,e,f,g,h,k,m,w,z,u,n,E,K)}catch(y){X(da);if(y!==y+0)throw y;Y(1,0)}}
-function Wb(){var a=Z;a=Object.assign({},a);var b=c=>d=>c(d)>>>0;a.V=b(a.V);a.W=b(a.W);a.ba=b(a.ba);a.ca=(c=>()=>c()>>>0)(a.ca);a.ga=b(a.ga);return a}
-function Xb(){function a(){l.calledRun=!0;if(!r){Z.U();l.onRuntimeInitialized?.();if(l.postRun)for("function"==typeof l.postRun&&(l.postRun=[l.postRun]);l.postRun.length;){var b=l.postRun.shift();ua.push(b)}ta(ua)}}if(0<G)H=Xb;else{if(l.preRun)for("function"==typeof l.preRun&&(l.preRun=[l.preRun]);l.preRun.length;)wa();ta(va);0<G?H=Xb:l.setStatus?(l.setStatus("Running..."),setTimeout(()=>{setTimeout(()=>l.setStatus(""),1);a()},1)):a()}}var Z;
-(async function(){function a(c){Z=c.exports;c=Z=Wb();xb=c.V;Xa=c.W;yb=c.Y;T=c.Z;Y=c._;L=c.$;X=c.aa;vb=c.ba;W=c.ca;zb=c.da;Ab=c.ea;Aa=c.fa;Bb=c.ga;F=c.T;Sa=c.X;la();G--;l.monitorRunDependencies?.(G);0==G&&H&&(c=H,H=null,c());return Z}G++;l.monitorRunDependencies?.(G);var b={a:Vb};if(l.instantiateWasm)return new Promise(c=>{l.instantiateWasm(b,(d,e)=>{c(a(d,e))})});na??=l.locateFile?l.locateFile("csv_converter.wasm",ca):ca+"csv_converter.wasm";return a((await qa(b)).instance)})();Xb();
+var xb,Xa,yb,T,Y,L,X,vb,W,zb,Ab,Aa,Bb,F,Sa,Xb={q:function(a){a>>>=0;var b=new za(a);0==t[b.ka+12>>>0]&&(t[b.ka+12>>>0]=1,ya--);t[b.ka+13>>>0]=0;xa.push(b);Ab(a);return Bb(a)},t:()=>{Y(0,0);var a=xa.pop();zb(a.qa);J=0},a:function(){return Ba([])},j:function(a){return Ba([a>>>0])},T:()=>{var a=xa.pop();a||ma("no exception to throw");var b=a.qa;0==t[a.ka+13>>>0]&&(xa.push(a),t[a.ka+13>>>0]=1,t[a.ka+12>>>0]=0,ya++);J=b;throw J;},k:function(a,b,c){a>>>=0;var d=new za(a);D[d.ka+16>>>2>>>0]=0;D[d.ka+4>>>
+2>>>0]=b>>>0;D[d.ka+8>>>2>>>0]=c>>>0;J=a;ya++;throw J;},S:()=>ya,c:function(a){J||=a>>>0;throw J;},O:()=>ma(""),D:function(a,b,c,d,e){a>>>=0;c>>>=0;b=M(b>>>0);d=0n===d;let f=g=>g;if(d){const g=8*c;f=h=>BigInt.asUintN(g,h);e=f(e)}Q(a,{name:b,ja:f,na:(g,h)=>{"number"==typeof h&&(h=BigInt(h));return h},oa:Fa(b,c,!d),ma:null})},G:function(a,b,c,d){b=M(b>>>0);Q(a>>>0,{name:b,ja:function(e){return!!e},na:function(e,f){return f?c:d},oa:function(e){return this.ja(v[e>>>0])},ma:null})},E:function(a){return Q(a>>>
+0,Ja)},C:function(a,b,c){b=M(b>>>0);Q(a>>>0,{name:b,ja:d=>d,na:(d,e)=>e,oa:Ka(b,c>>>0),ma:null})},s:function(a,b,c,d,e,f,g){a>>>=0;d>>>=0;e>>>=0;f>>>=0;var h=Pa(b,c>>>0);a=M(a);a=cb(a);e=Va(d,e);Oa(a,function(){ab(`Cannot call ${a} due to unbound types`,h)},b-1);bb(h,k=>{var m=a;var w=a;k=[k[0],null].concat(k.slice(1));var z=e,u=f,n=k.length;if(2>n)throw new P("argTypes array size mismatch! Must at least get return value and 'this' types!");var E=null!==k[1]&&!1,K=Ma(k),da=!k[0].sa,y=k[0],x=k[1];
+z=[w,Da,z,u,La,y.ja.bind(y),x?.na.bind(x)];for(u=2;u<n;++u)y=k[u],z.push(y.na.bind(y));if(!K)for(u=E?1:2;u<k.length;++u)null!==k[u].ma&&z.push(k[u].ma);K=Ma(k);u=k.length-2;x=[];y=["fn"];E&&y.push("thisWired");for(n=0;n<u;++n)x.push(`arg${n}`),y.push(`arg${n}Wired`);x=x.join(",");y=y.join(",");x=`return function (${x}) {\n`;K&&(x+="var destructors = [];\n");var Za=K?"destructors":"null",ra="humanName throwBindingError invoker fn runDestructors fromRetWire toClassParamWire".split(" ");E&&(x+=`var thisWired = toClassParamWire(${Za}, this);\n`);
+for(n=0;n<u;++n){var $a=`toArg${n}Wire`;x+=`var arg${n}Wired = ${$a}(${Za}, arg${n});\n`;ra.push($a)}x+=(da||g?"var rv = ":"")+`invoker(${y});\n`;if(K)x+="runDestructors(destructors);\n";else for(n=E?1:2;n<k.length;++n)E=1===n?"thisWired":"arg"+(n-2)+"Wired",null!==k[n].ma&&(x+=`${E}_dtor(${E});\n`,ra.push(`${E}_dtor`));da&&(x+="var ret = fromRetWire(rv);\nreturn ret;\n");k=(new Function(ra,x+"}\n"))(...z);w=Object.defineProperty(k,"name",{value:w});k=b-1;if(!l.hasOwnProperty(m))throw new Qa("Replacing nonexistent public symbol");
+void 0!==l[m].la&&void 0!==k?l[m].la[k]=w:(l[m]=w,l[m].pa=k);return[]})},r:function(a,b,c,d,e){a>>>=0;c>>>=0;b=M(b>>>0);let f=h=>h;if(0===d){var g=32-8*c;f=h=>h<<g>>>g;e=f(e)}Q(a,{name:b,ja:f,na:(h,k)=>k,oa:Fa(b,c,0!==d),ma:null})},m:function(a,b,c){function d(f){return new e(t.buffer,D[f+4>>>2>>>0],D[f>>>2>>>0])}var e=[Int8Array,Uint8Array,Int16Array,Uint16Array,Int32Array,Uint32Array,Float32Array,Float64Array,BigInt64Array,BigUint64Array][b];c=M(c>>>0);Q(a>>>0,{name:c,ja:d,oa:d},{ra:!0})},F:function(a,
+b){b=M(b>>>0);Q(a>>>0,{name:b,ja(c){var d=(d=c+4>>>0)?gb(d,D[c>>>2>>>0],!0):"";T(c);return d},na(c,d){d instanceof ArrayBuffer&&(d=new Uint8Array(d));var e="string"==typeof d;if(!(e||ArrayBuffer.isView(d)&&1==d.BYTES_PER_ELEMENT))throw new P("Cannot pass non-string to std::string");var f=e?db(d):d.length;var g=xb(4+f+1),h=g+4;D[g>>>2>>>0]=f;e?U(d,h,f+1):v.set(d,h>>>0);null!==c&&c.push(T,g);return g},oa:Ia,ma(c){T(c)}})},y:function(a,b,c){b>>>=0;c>>>=0;c=M(c);if(2===b){var d=ib;var e=jb;var f=kb}else d=
+lb,e=mb,f=nb;Q(a>>>0,{name:c,ja:g=>{var h=d(g+4,D[g>>>2>>>0]*b,!0);T(g);return h},na:(g,h)=>{if("string"!=typeof h)throw new P(`Cannot pass non-string to C++ string type ${c}`);var k=f(h),m=xb(4+k+b);D[m>>>2>>>0]=k/b;e(h,m+4,k+b);null!==g&&g.push(T,m);return m},oa:Ia,ma(g){T(g)}})},J:function(a,b){b=M(b>>>0);Q(a>>>0,{sa:!0,name:b,ja:()=>{},na:()=>{}})},M:()=>{I=!1;ob=0},N:(a,b)=>{V[a]&&(clearTimeout(V[a].id),delete V[a]);if(!b)return 0;var c=setTimeout(()=>{delete V[a];rb(()=>yb(a,performance.now()))},
+b);V[a]={id:c,ta:b};return 0},P:function(a,b,c,d){c>>>=0;d>>>=0;var e=(new Date).getFullYear(),f=(new Date(e,0,1)).getTimezoneOffset();e=(new Date(e,6,1)).getTimezoneOffset();D[a>>>0>>>2>>>0]=60*Math.max(f,e);C[b>>>0>>>2>>>0]=Number(f!=e);b=g=>{var h=Math.abs(g);return`UTC${0<=g?"-":"+"}${String(Math.floor(h/60)).padStart(2,"0")}${String(h%60).padStart(2,"0")}`};a=b(f);b=b(e);e<f?(U(a,c,17),U(b,d,17)):(U(a,d,17),U(b,c,17))},U:function(a){a>>>=0;var b=v.length;if(4294901760<a)return!1;for(var c=1;4>=
+c;c*=2){var d=b*(1+.2/c);d=Math.min(d,a+100663296);a:{d=(Math.min(4294901760,65536*Math.ceil(Math.max(a,d)/65536))-F.buffer.byteLength+65535)/65536|0;try{F.grow(d);la();var e=1;break a}catch(f){}e=void 0}if(e)return!0}return!1},Q:function(a,b){a>>>=0;b>>>=0;var c=0,d=0,e;for(e of ub()){var f=b+c;D[a+d>>>2>>>0]=f;c+=U(e,f,Infinity)+1;d+=4}return 0},R:function(a,b){a>>>=0;b>>>=0;var c=ub();D[a>>>2>>>0]=c.length;a=0;for(var d of c)a+=db(d)+1;D[b>>>2>>>0]=a;return 0},B:Cb,H:Db,b:Eb,v:Fb,d:Gb,g:Hb,z:Ib,
+K:Jb,p:Kb,l:Lb,A:Mb,x:Nb,I:Ob,f:Pb,n:Qb,e:Rb,h:Sb,i:Tb,o:Ub,u:Vb,w:Wb,L:qb};function Rb(a,b,c){var d=W();try{S(a)(b,c)}catch(e){X(d);if(e!==e+0)throw e;Y(1,0)}}function Pb(a){var b=W();try{S(a)()}catch(c){X(b);if(c!==c+0)throw c;Y(1,0)}}function Eb(a,b){var c=W();try{return S(a)(b)}catch(d){X(c);if(d!==d+0)throw d;Y(1,0)}}function Tb(a,b,c,d,e){var f=W();try{S(a)(b,c,d,e)}catch(g){X(f);if(g!==g+0)throw g;Y(1,0)}}
+function Hb(a,b,c,d){var e=W();try{return S(a)(b,c,d)}catch(f){X(e);if(f!==f+0)throw f;Y(1,0)}}function Gb(a,b,c){var d=W();try{return S(a)(b,c)}catch(e){X(d);if(e!==e+0)throw e;Y(1,0)}}function Sb(a,b,c,d){var e=W();try{S(a)(b,c,d)}catch(f){X(e);if(f!==f+0)throw f;Y(1,0)}}function Cb(a,b){var c=W();try{return S(a)(b)}catch(d){X(c);if(d!==d+0)throw d;Y(1,0)}}function Fb(a,b,c){var d=W();try{return S(a)(b,c)}catch(e){X(d);if(e!==e+0)throw e;Y(1,0)}}
+function Lb(a,b,c,d,e,f,g){var h=W();try{return S(a)(b,c,d,e,f,g)}catch(k){X(h);if(k!==k+0)throw k;Y(1,0)}}function Qb(a,b){var c=W();try{S(a)(b)}catch(d){X(c);if(d!==d+0)throw d;Y(1,0)}}function Kb(a,b,c,d,e,f){var g=W();try{return S(a)(b,c,d,e,f)}catch(h){X(g);if(h!==h+0)throw h;Y(1,0)}}function Ib(a,b,c,d,e){var f=W();try{return S(a)(b,c,d,e)}catch(g){X(f);if(g!==g+0)throw g;Y(1,0)}}function Jb(a,b,c,d,e,f){var g=W();try{return S(a)(b,c,d,e,f)}catch(h){X(g);if(h!==h+0)throw h;Y(1,0)}}
+function Ob(a,b,c){var d=W();try{return S(a)(b,c)}catch(e){X(d);if(e!==e+0)throw e;Y(1,0)}}function Mb(a,b,c,d,e,f,g,h){var k=W();try{return S(a)(b,c,d,e,f,g,h)}catch(m){X(k);if(m!==m+0)throw m;Y(1,0)}}function Ub(a,b,c,d,e,f,g,h){var k=W();try{S(a)(b,c,d,e,f,g,h)}catch(m){X(k);if(m!==m+0)throw m;Y(1,0)}}function Nb(a,b,c,d,e,f,g,h,k,m,w,z){var u=W();try{return S(a)(b,c,d,e,f,g,h,k,m,w,z)}catch(n){X(u);if(n!==n+0)throw n;Y(1,0)}}
+function Vb(a,b,c,d,e,f,g,h,k,m,w){var z=W();try{S(a)(b,c,d,e,f,g,h,k,m,w)}catch(u){X(z);if(u!==u+0)throw u;Y(1,0)}}function Wb(a,b,c,d,e,f,g,h,k,m,w,z,u,n,E,K){var da=W();try{S(a)(b,c,d,e,f,g,h,k,m,w,z,u,n,E,K)}catch(y){X(da);if(y!==y+0)throw y;Y(1,0)}}function Db(a,b,c){var d=W();try{return S(a)(b,c)}catch(e){X(d);if(e!==e+0)throw e;Y(1,0)}}function Yb(){var a=Z;a=Object.assign({},a);var b=c=>d=>c(d)>>>0;a.X=b(a.X);a.Y=b(a.Y);a.da=b(a.da);a.ea=(c=>()=>c()>>>0)(a.ea);a.ia=b(a.ia);return a}
+function Zb(){function a(){l.calledRun=!0;if(!r){Z.W();l.onRuntimeInitialized?.();if(l.postRun)for("function"==typeof l.postRun&&(l.postRun=[l.postRun]);l.postRun.length;){var b=l.postRun.shift();ua.push(b)}ta(ua)}}if(0<G)H=Zb;else{if(l.preRun)for("function"==typeof l.preRun&&(l.preRun=[l.preRun]);l.preRun.length;)wa();ta(va);0<G?H=Zb:l.setStatus?(l.setStatus("Running..."),setTimeout(()=>{setTimeout(()=>l.setStatus(""),1);a()},1)):a()}}var Z;
+(async function(){function a(c){Z=c.exports;c=Z=Yb();xb=c.X;Xa=c.Y;yb=c._;T=c.$;Y=c.aa;L=c.ba;X=c.ca;vb=c.da;W=c.ea;zb=c.fa;Ab=c.ga;Aa=c.ha;Bb=c.ia;F=c.V;Sa=c.Z;la();G--;l.monitorRunDependencies?.(G);0==G&&H&&(c=H,H=null,c());return Z}G++;l.monitorRunDependencies?.(G);var b={a:Xb};if(l.instantiateWasm)return new Promise(c=>{l.instantiateWasm(b,(d,e)=>{c(a(d,e))})});na??=l.locateFile?l.locateFile("csv_converter.wasm",ca):ca+"csv_converter.wasm";return a((await qa(b)).instance)})();Zb();
